+++ src/Checkbox/Checkbox.ts
@@ -1,7 +1,7 @@
 import { defineComponent, h, mergeProps, VNode } from 'vue'
 import { reactiveOmit } from '@vueuse/core'
 import { ElCheckboxGroup, ElCheckbox } from 'element-plus'
-import { useSelectConfig, useEmitValue } from '../composables/index'
+import { useDataConfig, useEmitValue } from '../composables/index'
 import { modelValueEmit } from '../utils/index'
 import props from './props'
 import type { ICheckboxProps } from './index'
@@ -11,7 +11,7 @@ export function createDefault<T>(
   component: T,
   className: string
 ): () => VNode {
-  const configKeys = useSelectConfig(props)
+  const configKeys = useDataConfig()
   const emitValue = useEmitValue()
   const config = reactiveOmit(props, 'data', 'config')
