+++ src/Form/FormItem.ts
@@ -10,7 +10,8 @@ import {
 } from 'vue'
 import { ElFormItem, ElButton } from 'element-plus'
 import { Plus, Minus } from '@element-plus/icons-vue'
-import { useFormItemBind, useFormChild, useCol } from '../composables/index'
+import { useCol } from '../composables/index'
+import { useFormItemBind, useFormChild } from './useForm'
 import { get, set, has, isArray } from '../utils/index'
 import ProFormItem from './FormItem'
 import ProFormComponent from './FormComponent'
