+++ src/InputTag/index.ts
@@ -4,6 +4,7 @@ import props from './props'
 import emits from './emits'
 import type { IDefineProps, IDefineEmits } from '../types/index'
 
+export * from './useInputTag'
 export const ProInputTag = withInstall(InputTag)
 export type IInputTagProps = IDefineProps<typeof props>
 export type IInputTagEmits = IDefineEmits<typeof emits>
